Rename date-filters to date-utils
The file now contains both filters and utility functions like
getNextMeetingDate, so the new name better reflects its contents.

- src/_lib/date-filters.js → src/_lib/date-utils.js
- tests/date-filters.test.js → tests/date-utils.test.js
- Update imports in .eleventy.js and test file
